Actualización del proyecto
Se juntó la interfaz con la parte de genereación de archivos.

+++ nbReman/src/org/harley/reman/reportes/xml/XMLConverter.java
@@ -182,15 +182,21 @@ public class XMLConverter {
      */
     public void convertXML2PDF(File xml, File xsl, File fo, File pdf){
         
-        //Convert from XML and XLS to FOP
-        convertXML2FO(xsl, xml, fo);
+        try{
+            //Convert from XML and XLS to FOP
+            convertXML2FO(xsl, xml, fo);
 
-        //Convert from FOP to PDF
-        convertFO2PDF(fo, pdf);
-        
-        //Delete temporal directory
-        //clean(fo, xml);
-        abrirPDF(pdf);
+            //Convert from FOP to PDF
+            convertFO2PDF(fo, pdf);
+
+            abrirPDF(pdf);
+        }
+        catch (Exception ex) {
+        }
+        finally {
+            //Delete temporal directory
+            clean(fo, xml);
+        }
     }
     
     /**
@@ -210,8 +216,10 @@ public class XMLConverter {
      * @param fo Archivo FO creado
      */
     public void clean(File fo, File xml){
-        fo.delete();
-        xml.delete();
+        if (fo.exists())
+            fo.delete();
+        if (xml.exists())
+            xml.delete();
     }
     
 }
